+++ packages/formik/src/FormSchema.tsx
@@ -1,7 +1,8 @@
-import { Schema } from '@domain-schema/core';
+import DomainSchema, { Schema } from '@domain-schema/core';
 import * as React from 'react';
 
 export default abstract class FormSchema extends Schema {
+  public id = DomainSchema.Int;
   public constructor() {
     super();
   }
