+++ packages/formik/src/fieldTypes.ts
@@ -1,8 +1,11 @@
+import { Button, RenderCheckBox, RenderField, RenderRadio, RenderSelect } from './components';
+import { FieldType, RenderComponentProps } from './types';
+
 export default {
-  custom: 'custom',
-  input: 'input',
-  select: 'select',
-  checkbox: 'checkbox',
-  radio: 'radio',
-  button: 'button'
+  custom: { name: 'custom' },
+  input: { name: 'input', component: RenderField },
+  select: { name: 'select', component: RenderSelect },
+  checkbox: { name: 'checkbox', component: RenderCheckBox },
+  radio: { name: 'radio', component: RenderCheckBox },
+  button: { name: 'button', component: Button }
 };
