+++ packages/formik/src/types.ts
@@ -3,7 +3,7 @@ import { ComponentType } from 'react';
 
 export interface SchemaField {
   type: any;
-  fieldType: string;
+  fieldType: FieldType;
   input?: any;
   defaultValue?: string | boolean | number;
   fieldAttrs?: any;
@@ -11,3 +11,15 @@ export interface SchemaField {
 }
 
 export type FSF = SchemaField & FieldValidators; // Full Schema Field
+
+export interface RenderComponentProps {
+  input?: any;
+  meta?: any;
+  children?: any[];
+  options?: any;
+}
+
+export interface FieldType {
+  name: string;
+  component?: ComponentType<any>;
+}
